Safari-609.1.11.1/LayoutTests/editing/selection/move-selection-backward-at-beginning-of-line-by-word-granularity.html

<!DOCTYPE html>
<html>
<body>
<div id="editor" contenteditable><div>first</div><div id="target">second</div></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests moving the selection backwards at the beginning of a line by word granularity');

editor.focus();
getSelection().setPosition(target, 0);
getSelection().modify('move', 'backward', 'word');

Markup.dump(editor);

</script>
</body>
</html>
