Safari-609.1.11.1/LayoutTests/editing/selection/move-selection-backward-at-beginning-of-line-by-word-granularity-expected.txt

This tests moving the selection backwards at the beginning of a line by word granularity
| <div>
|   "<#selection-caret>first"
| <div>
|   id="target"
|   "second"
